doctor-service/requirements.txt
django==5.0.4
djangorestframework==3.15.1
djangorestframework-simplejwt==5.3.1
psycopg2-binary==2.9.9
python-consul==1.1.0
django-cors-headers==4.3.1
gunicorn==21.2.0
python-dotenv==1.0.1
PyJWT==2.8.0
